+++ releasenotes.md
@@ -5,6 +5,7 @@
 ### Changes / improvements
 - `$typefrom` now also accepts a constant string, and so works like `$evaltype`.
 - `$evaltype` is deprecated in favour of `$typefrom`.
+- `-0xFF` will now be a signed integer.
 
 ### Fixes
 
@@ -71,6 +72,7 @@
 - Bug using `#foo` arguments with `$defined` #2173
 - Incorrect ensure on String.split.
 - Removed the naive check for compile time modification, which fixes #1997 but regresses in detection.
+- `-2147483648`, MIN literals work correctly.
 
 ### Stdlib changes
 - Added `String.quick_ztr` and `String.is_zstr`